// client/src/schemas/heroSchemas.ts
import { z } from 'zod';

export const heroSchemas = z.object({
  id: z.number(),
  name: z.string(),
  status: z.string(),
  species: z.string(),
  type: z.string(),
  gender: z.string(),
  origin: z.object({
    name: z.string(),
    url: z.string(),
  }),
  location: z.object({
    name: z.string(),
    url: z.string(),
  }),
  image: z.string(),
  created: z.string(),
});

export type CardType = z.infer<typeof heroSchemas>;